refactor: extract prompt components
