+++ src/app/(main)/project/[id]/_components/view/ItemBlock.tsx
@@ -1,0 +1,7 @@
+import type { ReactNode } from "react";
+
+export function ItemBlock({ children }: { children: ReactNode }) {
+	return (
+		<div className="border-2 border-slate-300 p-1 rounded">{children}</div>
+	);
+}
