+++ src/app/(main)/project/[id]/_components/view/TryItView.tsx
@@ -1,0 +1,8 @@
+"use client";
+export function TryItView() {
+	return (
+		<div className="md:col-span-2 h-[40rem] bg-slate-700 p-2">
+			<h4 className="font-bold text-lg text-white">Try It</h4>
+		</div>
+	);
+}
